3.1-drf-intro/smart_home/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
generics==6.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
